backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.12.0
pdfplumber==0.11.0
regex==2023.12.25
google-generativeai==0.3.2
groq==0.4.1
reportlab==4.0.7
